Fix Combustion Heater deleting remaining items > 64
Fixes an issue where, for example, putting five stacks of charcoal into
the chamber and activating the heater will result in the crafted amount
of coal being placed in the collector, and only a single stack of
charcoal remaining in the chamber instead of the full remainder.

This change causes the remainder amount to spawn several stacks
according to the item's stack limit.